Format the sql queries when the dbix class debug flag is set
Rather than all on one line this splits them the query over a few lines to make
it easier to read.

We also (very crudely) attempt to swap the query variables into the query by
regex'ing out the placeholders.  In general that's bad, but for debugging it
makes it easy to copy & paste the query into a sql command line to work on.