Do not strip whitespace & non alpha numeric chars
We were doing this originally as the only use for these columns was to building
full text indexes for searching.

But I now want to use the data for display in search results, so leaving it
as-is makes more sense.  This will increase the data size some, and likely
making indexing a little longer, but neither should be dramatic or of any
noticeable concern.